refactor: 既存フックを presentation/hooks/ に移動、PuzzleBoard を更新
## フック移動
- useKeyboard, useSwipe, useHintMode, useCompletionOverlay,
  useVideoPlayback, usePuzzleTimer を presentation/hooks/ にコピー
- 新パスのインポートを修正（ui-atoms, game-atoms, puzzle-atoms を直接参照）
- 旧 hooks/ は後方互換の再エクスポートに変換

## PuzzleBoard 更新
- フックインポートを presentation/hooks/ に変更
- PuzzleBoard.test.tsx のモックパスを更新

全 6419 テストパス、ビルド成功
