+++ src/hooks/useCompletionOverlay.ts
@@ -1,39 +1,2 @@
-import { useAtom } from 'jotai';
-import { useCallback } from 'react';
-import { completionOverlayVisibleAtom } from '../presentation/store/ui-atoms';
-
-/**
- * 完成オーバーレイの表示/非表示を管理するカスタムフック
- */
-export const useCompletionOverlay = () => {
-  // 状態
-  const [overlayVisible, setOverlayVisible] = useAtom(completionOverlayVisibleAtom);
-
-  /**
-   * オーバーレイの表示状態を切り替える
-   */
-  const toggleOverlay = useCallback(() => {
-    setOverlayVisible(prev => !prev);
-  }, [setOverlayVisible]);
-
-  /**
-   * オーバーレイを表示する
-   */
-  const showOverlay = useCallback(() => {
-    setOverlayVisible(true);
-  }, [setOverlayVisible]);
-
-  /**
-   * オーバーレイを非表示にする
-   */
-  const hideOverlay = useCallback(() => {
-    setOverlayVisible(false);
-  }, [setOverlayVisible]);
-
-  return {
-    overlayVisible,
-    toggleOverlay,
-    showOverlay,
-    hideOverlay,
-  };
-};
+// 後方互換: presentation/hooks/useCompletionOverlay.ts に移動済み
+export { useCompletionOverlay } from '../presentation/hooks/useCompletionOverlay';
